method moved from one class into another class
